Lactate Threshold - Fix error of missing sport parameter (PR #120)
* Correction to Lactate Threshold - Speed

Speed requires a sport to function properly. RUNNING is used as the default as that is what Garmin currently allows that I can verify. SWIMMING and CYCLING can be used in the future if allowed by GARMIN.

`LACTATE_THRESHOLD_SPORTS` has a default value of `RUNNING` but additional values can be added **separated by comma and  without any space** if/when they are supported like `RUNNING,CYCLING`